Add player information GUI (position and rotation). Removed some TODOs.

+++ src/particles/particleVShader.glsl
@@ -11,13 +11,13 @@ out vec2 textureCoords2;  // next texture atlas stage
 out float blend;          // blend factor between stages
 
 uniform mat4 projectionMatrix;
-uniform float numberOfRows;
+uniform int numberOfRows;
 
 void main(void){
 
 	vec2 textureCoords = position + vec2(0.5, 0.5);
 	textureCoords.y = 1.0 - textureCoords.y; // texture coordinate system has y axis inverted
-	textureCoords /= numberOfRows;
+	textureCoords /= float(numberOfRows);
 	textureCoords1 = textureCoords + texOffsets.xy;
 	textureCoords2 = textureCoords + texOffsets.zw;
 	blend = blendFactor;
